Created IRosieEntity and Added db operation methods to DBConnector class

+++ AuntRosieEntities/Nutrition.cs
@@ -7,7 +7,7 @@ namespace AuntRosieEntities
 {
     public class Nutrition
     {
-        private int id;
+        private short id;
         private string name;
 
         public Nutrition Add()
